Updated cluster annotations for immune cell clusters
For many immune cell clusters, the sample contribution from decidual and villous samples is about equal. This could suggest that the same immune cell types exist on both sides of the fetal-maternal interface and they end up in the same cluster because they are the same cell type. Initially we have labelled most of them as decidual (with a prefix "dec."), but to account for this uncertainty about their tissue of origin, we are dropping the prefix for the immune cell clusters (except vil.Hofb because the source is implied).