Update files on GEO/SRA
We realized that one of the samples (CTRL_2, villi) was sequenced twice
and both sets of files were used for the analysis, but during submission
to NCBI GEO, we missed uploading one set of the files. Changes in this
commit are for redownloading all files, putting together their metadata,
etc. for submission to GEO/SRA for updating the data files there.